requirements.txt
astroid==1.6.1
autopep8==1.3.4
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.7
django-allauth==0.36.0
django-angular==2.0.2
django-autoslug-field==0.2.3
django-heroku==0.3.1
django-toolbelt==0.0.1
djangorestframework==3.8.2
factory-boy==2.10.0
Faker==0.8.12
gunicorn==19.7.1
idna==2.6
isort==4.3.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
numpy==1.14.5
oauthlib==2.1.0
pandas==0.23.3
pep8==1.7.1
pipenv==2018.7.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycodestyle==2.3.1
pylint==1.8.2
PySocks==1.6.8
python-dateutil==2.7.2
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==1.0.0
six==1.11.0
static3==0.7.0
text-unidecode==1.2
tweepy==3.6.0
urllib3==1.22
uuid==1.30
virtualenv==15.1.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
wrapt==1.10.11
